move new version to similar_regions module and update docs and examples
